pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "openai>=1.61.1",
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "telebot>=0.0.5",
]
